+++ src/main/java/com/goormitrip/goormitrip/repository/UserRepository.java
@@ -1,11 +1,12 @@
 package com.goormitrip.goormitrip.repository;
 
-import com.goormitrip.goormitrip.domain.User;
+import com.goormitrip.goormitrip.domain.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
     boolean existsByEmail(String email);
-    Optional<User> findByEmail(String email);
+
+    Optional<UserEntity> findByEmail(String email);
 }
